fix: restructurisation of Group
