+++ src/components/Loading.js
@@ -1,0 +1,12 @@
+import React from 'react'
+import Spinner from 'react-bootstrap/Spinner'
+
+function Loading() {
+    return (
+        <Spinner animation="border" role="status">
+            <span className="visually-hidden">Loading...</span>
+        </Spinner>
+    )
+}
+
+export default Loading
